Improve coverage
