+++ analyser/application/markdown/tests/test_markdown.py
@@ -1,9 +1,5 @@
-from unittest.mock import MagicMock, patch
-
-from application.catalogued_repository import CataloguedRepository
-from application.markdown.markdown import create_markdown_file, set_up_index_page, set_up_markdown_file
-
-FILE_PATH = "application.markdown.markdown"
+from unittest.mock import MagicMock
+from application.markdown.markdown import create_markdown_file, set_up_markdown_file
 
 
 def test_set_up_markdown_file() -> None:
@@ -24,20 +20,3 @@ def test_create_markdown_file() -> None:
     create_markdown_file(markdown_file)
     # Assert
     markdown_file.create_md_file.assert_called_once()
-
-
-@patch(f"{FILE_PATH}.set_up_markdown_file")
-@patch(f"{FILE_PATH}.create_markdown_file")
-def test_set_up_index_page(
-    mock_create_markdown_file: MagicMock,
-    mock_set_up_markdown_file: MagicMock,
-) -> None:
-    # Arrange
-    repositories = [
-        CataloguedRepository("JackPlowman/github-stats", "A repository for analysing GitHub repositories.", 100, 10)
-    ]
-    # Act
-    set_up_index_page(repositories)
-    # Assert
-    mock_set_up_markdown_file.assert_called_once_with("index", "Active Repository Statistics")
-    mock_create_markdown_file.assert_called_once_with(mock_set_up_markdown_file.return_value)
